+++ src/styles/globals.scss
@@ -180,7 +180,7 @@
 
 body[color-theme='light'] {
   //System Color
-  --system_background: #{$-Success-07};
+  --system_background: #{$-White};
   --system_ui-01: #{$-Gray-01};
   --system_ui-02: #{$-Gray-02};
   --system_ui-03: #{$-Gray-03};
