+++ src/styles/room-list.module.scss
@@ -9,8 +9,8 @@
   margin: 0 auto;
   width: 1040px;
   height: 700px;
-  background: $-system_white;
-  border: 0.5px solid $-gray_cool-200;
+  background: var(--system_background);
+  border: 0.5px solid var(--system_ui-04);
   border-radius: 16px;
 
   .room-list-info{
@@ -31,7 +31,7 @@
     margin-top: 20px;
     margin-left: 6px;
     font-weight: 700;
-    color: $-gray_cool-800;
+    color: var(--text-01);
   }
   .room-scroll {
     height: 632px;
@@ -41,7 +41,7 @@
 
     .room-scroll-text{
       text-align: center;
-      color: $-gray_cool-400;
+      color: var(--system_ui-05);
     }
 
     .room-list-grid {
@@ -52,8 +52,8 @@
 
       .room-list-form {
         height: 128px;
-        background: $-system_white;
-        border: 0.5px solid $-gray_cool-200;
+        background: var(--ui-cardui);
+        border: 0.5px solid var(--system_ui-04);
         border-radius: 12px;
         display: inline-flex;
         margin-bottom: 20px;
@@ -64,13 +64,14 @@
           left: 16px;
           top: 16px;
           margin: 16px;
-          background: $-gray_cool-100;
+          background: var(--system_ui-02);
           border-radius: 12px;
         }
 
         .room-title {
           margin: 16px 12px 0px 0px;
           height: 28px;
+          color: var(--text-03);
           font-weight: 700;
         }
 
@@ -79,7 +80,7 @@
           height: 22px;
           margin-top: 4px;
           font-weight: 400;
-          color: $-gray_cool-400;
+          color: var(--system_ui-05);
         }
 
         .room-joiner-img-blank {
@@ -87,7 +88,7 @@
           width: 32px;
           margin-top: 8px;
           border: none;
-          background: $-gray_cool-50;
+          background: var(--system_ui-01);
           border-radius: 8px;
           margin-right: 8px;
         }
@@ -105,7 +106,7 @@
           width: 86px;
           height: 38px;
           margin-top: 7px;
-          background: $-primary-500;
+          background: var(--primary-press);
           border: none;
           border-radius: 8px;
           display: flex;
@@ -117,7 +118,7 @@
             font-weight: 500;
             font-size: 16px;
             line-height: 22px;
-            color: $-system_white;
+            color: var(--primary);
             user-select: none;
             cursor: default;
           }
@@ -127,7 +128,7 @@
           }
 
           &:active {
-            background: $-primary-600;
+            background: var(--primary-press);
           }
 
           &:disabled {
